Improved API publishing payload
Previously in the API_dictionary, API name was used as the key and
when a user deletes and API, API name was passed to delete it. This
might causes errors if there is another API with same name exists in
another organisation or a tenant. API ID is used instead of the name.

Signed-off-by: Lakshitha Gunasekara <lakshitha.gunasekara@gmail.com>